requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
feedparser
beautifulsoup4
requests
pydantic-settings
ollama
python-dateutil